# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Event Management API",
    "version": "1.0.0",
    "description": "API documentation for the Event Management backend"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Venue": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "668f1a1234567890abcdef12"
          },
          "name": {
            "type": "string",
            "example": "Royal Lounge"
          },
          "type": {
            "type": "string",
            "enum": [
              "lounge",
              "theatre"
            ],
            "example": "lounge"
          },
          "address": {
            "type": "string",
            "example": "Main Street 123"
          },
          "image": {
            "type": "string",
            "example": "https://example.com/lounge.jpg"
          },
          "baseCost": {
            "type": "number",
            "example": 1500
          },
          "maxCapacity": {
            "type": "integer",
            "nullable": true,
            "example": 200
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/services": {
      "get": {
        "summary": "Get all services (hospitality,Camera,MusicalBand....etc)",
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Services configuration loaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ServicesConfig"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/services/venues": {
      "get": {
        "summary": "Get all active venues ",
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "lounge",
                "theatre"
              ]
            },
            "description": "Filter by venue type"
          }
        ],
        "responses": {
          "200": {
            "description": "List of active venues",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Venue"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  }
}